Move python modules outside of package; now they're importable
