+++ setup.py
@@ -10,7 +10,6 @@ def readme():
         return f.read()
 
 
-
 setup(name='thinkx',
       version='1.1',
       description='Modules supporting books by Allen Downey',
