+++ tests/thinkdsp_test.py
@@ -8,7 +8,7 @@ License: GNU GPLv3 http://www.gnu.org/licenses/gpl.html
 from __future__ import print_function, division
 
 import unittest
-from thinkx import thinkdsp
+import thinkdsp
 
 import numpy as np
