+++ thinkbayes.py
