+++ thinkbayes2.py
