+++ thinkplot.py
