+++ thinkstats2.py
